🎉 feat: Add MousePosition struct and tests 📝 A new module named `mouse` was created in the `src` directory. The module contains a `MousePosition` struct which represents the current position of the mouse on the screen.
The `MousePosition` struct has two public fields `x` and `y` which are integers representing the horizontal and vertical coordinates of the mouse respectively. A `new` function was implemented to create a new instance of the `MousePosition` struct.

In addition, a test module was added to the `mouse.rs` file. The test module contains a unit test for the `new` function, which checks that a new instance of `MousePosition` is created correctly with the provided coordinates.

Changes made:
- Created `src/mouse.rs` file and implemented `MousePosition` struct.
- Implemented `new` function for `MousePosition` struct.
- Added a `tests` module to the `mouse.rs` file to test the `new` function.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>