🔍 review(src/mouse.rs): Implement the `offset` method for MousePosition
A new method `offset` was added to the `MousePosition` struct which allows the user to get a new `MousePosition` object with the current position offset by a given distance. This is useful when performing mouse actions that require movement in a certain direction.
This method takes two parameters, `distance_x` and `distance_y`, which are offsets from the current position of the mouse. The `x` and `y` fields of `MousePosition` are then adjusted by adding these distances.

Changes made:
- Added `offset` method which takes two parameters `distance_x` and `distance_y` to specify the offset distance in the x and y direction respectively.
- The method returns a new `MousePosition` object with the offset applied to the current position.

No issues were found in this code change. The code looks good and the test confirms the correct implementation of the `offset` method.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>